chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-script-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE ERROR: line 33: Uncaught TypeError: Cannot read property 'createElement' of undefined
Test that scripts in imported documents are executed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: Cannot read property 'createElement' of undefined
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
